Cleaning thing up
