+++ pkg/skaffold/docker/build_args_test.go
@@ -1,0 +1,109 @@
+/*
+Copyright 2019 The Skaffold Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package docker
+
+import (
+	"testing"
+
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest"
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/util"
+	"github.com/GoogleContainerTools/skaffold/testutil"
+)
+
+func TestGetBuildArgs(t *testing.T) {
+	tests := []struct {
+		description string
+		artifact    *latest.DockerArtifact
+		env         []string
+		want        []string
+		shouldErr   bool
+	}{
+		{
+			description: "build args",
+			artifact: &latest.DockerArtifact{
+				BuildArgs: map[string]*string{
+					"key1": util.StringPtr("value1"),
+					"key2": nil,
+					"key3": util.StringPtr("{{.FOO}}"),
+				},
+			},
+			env:  []string{"FOO=bar"},
+			want: []string{"--build-arg", "key1=value1", "--build-arg", "key2", "--build-arg", "key3=bar"},
+		},
+		{
+			description: "invalid build arg",
+			artifact: &latest.DockerArtifact{
+				BuildArgs: map[string]*string{
+					"key": util.StringPtr("{{INVALID"),
+				},
+			},
+			shouldErr: true,
+		},
+		{
+			description: "cache from",
+			artifact: &latest.DockerArtifact{
+				CacheFrom: []string{"gcr.io/foo/bar", "baz:latest"},
+			},
+			want: []string{"--cache-from", "gcr.io/foo/bar", "--cache-from", "baz:latest"},
+		},
+		{
+			description: "target",
+			artifact: &latest.DockerArtifact{
+				Target: "stage1",
+			},
+			want: []string{"--target", "stage1"},
+		},
+		{
+			description: "network mode",
+			artifact: &latest.DockerArtifact{
+				NetworkMode: "Bridge",
+			},
+			want: []string{"--network", "bridge"},
+		},
+		{
+			description: "no-cache",
+			artifact: &latest.DockerArtifact{
+				NoCache: true,
+			},
+			want: []string{"--no-cache"},
+		},
+		{
+			description: "all",
+			artifact: &latest.DockerArtifact{
+				BuildArgs: map[string]*string{
+					"key1": util.StringPtr("value1"),
+				},
+				CacheFrom:   []string{"foo"},
+				Target:      "stage1",
+				NetworkMode: "None",
+			},
+			want: []string{"--build-arg", "key1=value1", "--cache-from", "foo", "--target", "stage1", "--network", "none"},
+		},
+	}
+	for _, test := range tests {
+		testutil.Run(t, test.description, func(t *testutil.T) {
+			t.Override(&util.OSEnviron, func() []string { return test.env })
+
+			result, err := GetBuildArgs(test.artifact)
+
+			t.CheckError(test.shouldErr, err)
+			if !test.shouldErr {
+				t.CheckDeepEqual(test.want, result)
+			}
+		})
+	}
+}
